feat(release-pr): do not include the crate name if no workspace
