+++ crates/release_plz_core/src/release_pr.rs
@@ -1,16 +1,13 @@
 use std::path::PathBuf;
 
-use cargo_metadata::Package;
-use chrono::SecondsFormat;
 use git_cmd::Repo;
 
 use anyhow::{anyhow, Context};
 use tracing::instrument;
 
-use crate::backend::{GitClient, Pr};
-use crate::{copy_to_temp_dir, update, GitBackend, UpdateRequest, UpdateResult, CARGO_TOML};
-
-const BRANCH_PREFIX: &str = "release-plz/";
+use crate::backend::{GitClient};
+use crate::pr::{BRANCH_PREFIX, Pr};
+use crate::{copy_to_temp_dir, update, GitBackend, UpdateRequest, CARGO_TOML};
 
 #[derive(Debug)]
 pub struct ReleasePrRequest {
@@ -45,62 +42,13 @@ pub async fn release_pr(input: &ReleasePrRequest) -> anyhow::Result<()> {
         let repo = Repo::new(new_manifest_dir)?;
         let there_are_commits_to_push = repo.is_clean().is_err();
         if there_are_commits_to_push {
-            let pr = Pr::from((repo.default_branch(), packages_to_update.as_ref()));
+            let pr = Pr::new(repo.default_branch(), packages_to_update.as_ref() );
             create_release_branch(&repo, &pr.branch)?;
             gh_client.open_pr(&pr).await?;
         }
     }
 
     Ok(())
-}
-
-impl From<(&str, &[(Package, UpdateResult)])> for Pr {
-    fn from((default_branch, packages_to_update): (&str, &[(Package, UpdateResult)])) -> Self {
-        Self {
-            branch: release_branch(),
-            base_branch: default_branch.to_string(),
-            title: pr_title(packages_to_update),
-            body: pr_body(packages_to_update),
-        }
-    }
-}
-
-fn release_branch() -> String {
-    let now = chrono::offset::Utc::now();
-    // Convert to a string of format "2018-01-26T18:30:09Z".
-    let now = now.to_rfc3339_opts(SecondsFormat::Secs, true);
-    // ':' is not a valid character for a branch name.
-    let now = now.replace(':', "-");
-    format!("{BRANCH_PREFIX}{now}")
-}
-
-fn pr_title(packages_to_update: &[(Package, UpdateResult)]) -> String {
-    if packages_to_update.len() == 1 {
-        let (package, update) = &packages_to_update[0];
-        format!("chore({}): release v{}", package.name, update.version)
-    } else {
-        "chore: release".to_string()
-    }
-}
-
-fn pr_body(packages_to_update: &[(Package, UpdateResult)]) -> String {
-    let header = "## 🤖 New release";
-    let updates: String = packages_to_update
-        .iter()
-        .map(|(package, update)| {
-            if package.version != update.version {
-                format!(
-                    "\n* `{}`: {} -> {}",
-                    package.name, package.version, update.version
-                )
-            } else {
-                format!("\n* `{}`: {}", package.name, package.version)
-            }
-        })
-        .collect();
-    let footer =
-        "---\nThis PR was generated with [release-plz](https://github.com/MarcoIeni/release-plz/).";
-    format!("{header}{updates}\n{footer}")
 }
 
 fn create_release_branch(repository: &Repo, release_branch: &str) -> anyhow::Result<()> {
